fix: refactored menu out of Card component
